ARM: lib: pbl: add shared architected timer
To share the previously AARCH64 only implementation the delay code is
moved from the lib64 folder to the shared lib folder. Note that only
some ARMv7-A cores implement the 64bit wide generic timer CNTPCT[1] as
it is optional.

As the functionality must be opt-in it is guarded behind the PBL_TIMER
symbol which is enabled by default for AARCH64 if PBL_CLOCKSOURCE is not
implemented.

[1]: See "ARM Architecture Reference Manual ARMv7-A and ARMv7-R edition
(rev C.d)", Chapter B8 "The Generic Timer"

Signed-off-by: Stefan Kerkmann <s.kerkmann@pengutronix.de>