filetype: introduce filetype_fit
It's not possible to differentiate a flattened device tree from a FIT
image without parsing it for top-level /images and /configurations node.

We thus have some checks that check for filetype_oftree instead of
filetype_fit, which is confusing. Let's introduce filetype_fit and
handle it at the boot layer to allow for slightly more readable code.

This will pay off some more when we start making use of it in a later
commit.

Signed-off-by: Ahmad Fatoum <a.fatoum@barebox.org>